Try to the lb and nfs mistaken identity issue with a block
"Mistaken Identity" being that casl assumes anything in OSEv3 is a node
and tries to execute node-specific tasks. In openshift-ansible, lb and
nfs groups are placed in OSEv3, but don't have node-specific tasks run
against them, so we're going to try and fix that in casl for the HA
cases.